Move getRenderedProps to browser.js...
Because we don't need it in SSR, and can
cleanup/simplify/reduce-size-of app.js a little bit.
Tiny related refactoring.
