+++ src/encoded/static/browser.js
@@ -26,7 +26,7 @@ function updateSessionInfo(){
     // keep <script data-prop-type="user_details"> in <head>, before <script data-prop-type="inline">, in app.js
     // so is available before this JS (via bundle.js)
     try {
-        props = App.getRenderedPropValues(document, 'user_info');
+        props = getRenderedPropValues(document, 'user_info');
     } catch(e) {
         console.error(e);
         return false;
@@ -56,27 +56,66 @@ if (typeof window !== 'undefined' && window.document && !window.TEST_RUNNER) {
 
         // Update Redux store from Redux store props that've been rendered server-side
         // into <script data-prop-name={ propName }> elements.
-        var initialReduxStoreState = App.getRenderedProps(document);
+        const initialReduxStoreState = getRenderedProps(document);
         delete initialReduxStoreState.user_details; // Stored into localStorage.
         store.dispatch({ 'type' : initialReduxStoreState });
 
         const AppWithReduxProps = connect(mapStateToProps)(App);
-        let app;
 
         try {
-            app = ReactDOM.hydrate(<Provider store={store}><AppWithReduxProps /></Provider>, document);
+            ReactDOM.hydrate(<Provider store={store}><AppWithReduxProps /></Provider>, document);
         } catch (e) {
             console.error("INVARIANT ERROR", e); // To debug
             // So we can get printout and compare diff of renders.
-            app = require('react-dom/server').renderToString(<Provider store={store}><AppWithReduxProps /></Provider>);
+            window.app = require('react-dom/server').renderToString(<Provider store={store}><AppWithReduxProps /></Provider>);
         }
 
         // Set <html> class depending on browser features
         BrowserFeat.setHtmlFeatClass();
-
-        // Simplify debugging
-        window.app = app;
-        window.React = React;
     });
 
 }
+
+
+
+/**
+ * Collects prop values from server-side-rendered HTML
+ * to be re-fed into Redux store.
+ *
+ * @param {HTMLElement} document - HTML DOM element representing the document.
+ * @param {string} [filter=null] - If set, filters down prop fields/values collected to only one(s) defined.
+ * @returns {Object} Object keyed by field name with collected value as value.
+ */
+function getRenderedPropValues(document, filter = null){
+    const returnObj = {};
+    let script_props;
+    if (typeof filter === 'string'){
+        script_props = document.querySelectorAll('script[data-prop-name="' + filter + '"]');
+    } else {
+        script_props = document.querySelectorAll('script[data-prop-name]');
+    }
+    script_props.forEach(function(elem){
+        const prop_name = elem.getAttribute('data-prop-name');
+        let elem_value = elem.text;
+        const elem_type = elem.getAttribute('type') || '';
+        if (elem_type === 'application/json' || elem_type.slice(-5) === '+json') {
+            elem_value = JSON.parse(elem_value);
+        }
+        returnObj[prop_name] = elem_value;
+    });
+    return returnObj;
+}
+
+/**
+ * Runs `App.getRenderedPropValues` and adds `href` key value from canonical link element.
+ *
+ * @param {HTMLElement} document - HTML DOM element representing the document.
+ * @param {string} [filter=null] - If set, filters down prop fields/values collected to only one(s) defined.
+ * @returns {Object} Object keyed by field name with collected value as value.
+ */
+function getRenderedProps(document, filter = null) {
+    const returnObj = getRenderedPropValues(document, filter);
+    returnObj.href = document.querySelector('link[rel="canonical"]').getAttribute('href'); // Ensure the initial render is exactly the same
+    return returnObj;
+}
+
